gophercises/task/db/tasks.go
package db

import (
	"encoding/binary"
	"time"

	"github.com/boltdb/bolt"
)

var taskBucket = []byte("tasks")
var db *bolt.DB
var NewCreateTask = CreateTask
var NewDeleteTask = DeleteTask
var NewAllTasks = AllTasks

var newDbView = dbView
var newDbUpdate = dbUpdate

// Task is a struct which defines key value parameters
type Task struct {
	Key   int
	Value string
}

// Init opens db and create taskbucket if it is not exists.
func Init(dbPath string) error {
	var err error
	db, err = bolt.Open(dbPath, 0600, &bolt.Options{Timeout: 1 * time.Second})
	if err != nil {
		return err
	}
	return db.Update(func(tx *bolt.Tx) error {
		_, err := tx.CreateBucketIfNotExists(taskBucket)
		return err
	})
}

func dbUpdate(task string) (int, error) {
	var id int
	err := db.Update(func(tx *bolt.Tx) error {
		b := tx.Bucket(taskBucket)
		id64, _ := b.NextSequence()
		id = int(id64)
		key := itob(id)
		return b.Put(key, []byte(task))
	})
	return id, err
}

// CreateTask creates new task
func CreateTask(task string) (int, error) {
	id, err := newDbUpdate(task)
	if err != nil {
		return -1, err
	}
	return id, nil
}

func dbView() ([]Task, error) {
	var tasks []Task
	err := db.View(func(tx *bolt.Tx) error {
		b := tx.Bucket(taskBucket)
		c := b.Cursor()
		for k, v := c.First(); k != nil; k, v = c.Next() {
			tasks = append(tasks, Task{
				Key:   btoi(k),
				Value: string(v),
			})
		}
		return nil
	})
	return tasks, err
}

// AllTasks returns all tasks
func AllTasks() ([]Task, error) {
	tasks, err := newDbView()
	if err != nil {
		return nil, err
	}
	return tasks, nil
}

// DeleteTask deletes task of given key
func DeleteTask(key int) error {
	return db.Update(func(tx *bolt.Tx) error {
		b := tx.Bucket(taskBucket)
		return b.Delete(itob(key))
	})
}

func itob(v int) []byte {
	b := make([]byte, 8)
	binary.BigEndian.PutUint64(b, uint64(v))
	return b
}

func btoi(b []byte) int {
	return int(binary.BigEndian.Uint64(b))
}
